+++ CHANGELOG.md
@@ -1,9 +1,18 @@
 ## [Unreleased]
 
--   Fixed an issue where the absolute value key from the virtual keyboard would
+### Improvements
+
+-   The Symbols keyboard is now a top-level keyboard. Previously it was accessible
+    only from the Roman keyboard.
+
+### Bug Fixes
+
+-   The absolute value character "|" (and other small delimiters) would be displayed in the wrong font (and too small)
+
+-   The absolute value key from the virtual keyboard would
     insert '|#@|'
 
--   Fixed an issue where 'sqrt' from the virtual keyboard or keyboard shortcut
+-   The 'sqrt' key from the virtual keyboard or keyboard shortcut
     (option+V) would do nothing. The problem affected any inline shortcut or key
     that included a '#0' argument when there was no selection.